edtech_site/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-static==0.0.6
Django==4.1.3
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-utils-six==2.0
idna==3.4
Pillow==9.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
static-ranges==0.2.0
static3==0.7.0
stripe==5.5.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
